--- src/components/AboutProject/AboutProjectContainer.css
@@ -1,130 +1,0 @@
-.about-wrapper {
-    position: relative;
-    width: 100%;
-    height: 100%;
-    padding: 20px 15px;
-    overflow: hidden;
-    box-sizing: border-box;
-    background: #00007c;
-}
-
-.add-project {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    position: absolute;
-    right: 5px;
-    top: 5px;
-    padding: 7px;
-    background: whitesmoke;
-    font-size: 10px;
-    font-weight: bold;
-    border-radius: 15px;
-    box-shadow: 0 5px 7px 0 rgba(0,0,0,0.2);
-    transition: .3s;
-    cursor: pointer;
-}
-
-.add-project:hover {
-    box-shadow: 0 5px 7px 0 rgba(0,0,0,0.6);
-}
-
-.add-project img {
-    width: 10px;
-}
-
-.logo-info-item {
-    margin: 10px 0;
-}
-
-.logo-name {
-    font-size: 22px;
-}
-
-.logo-description {
-    font-size: 14px;
-    line-height: 1.5;
-    word-break: break-all;
-}
-
-.logo-created {
-    display: flex;
-}
-
-.logo-created input {
-    flex: 1;
-    width: auto;
-    margin-left: 10px;
-}
-
-.logo-website {
-    text-decoration: none;
-    opacity: .7;
-    transition: .2s;
-}
-
-.logo-website:hover {
-    opacity: 1;
-}
-
-.logo-info-input {
-    display: block;
-    width: 100%;
-    background: none;
-    box-sizing: border-box;
-    word-break: break-all;
-    border: none;
-}
-
-.logo-info-input:focus {
-    border-bottom: 1px solid #FB4747;
-}
-
-.color-picker-container {
-    display: flex;
-    justify-content: flex-start;
-}
-
-.bg-color-picker {
-    position: relative;
-    width: 20px;
-    height: 10px;
-    margin-right: 10px;
-    padding: 1px;
-    background: #fafafa;
-    border: 1px solid gray;
-    cursor: pointer;
-}
-
-.sketch-picker-container {
-    position: absolute;
-    top: 100%;
-}
-
-.info-wrapper {
-    height: 100%;
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-}
-
-.info-wrapper.enter-done {
-    opacity: 1;
-    transform: translateX(0);
-}
-.info-wrapper.enter-active {
-    opacity: 0;
-    transform: translateX(-40%);
-}
-.info-wrapper.exit {
-    opacity: 1;
-    transform: translateX(0%);
-}
-.info-wrapper.exit-active {
-    opacity: 0;
-    transform: translateY(-50%);
-}
-.info-wrapper.enter-done,
-.info-wrapper.exit-active {
-    transition: opacity 350ms, transform 350ms cubic-bezier(0.390, 0.575, 0.565, 1.000);
-}
