--- src/components/LeftToolbar/LeftToolbar.css
@@ -1,121 +1,0 @@
-.projects-wrapper {
-    position: relative;
-    height: 100vh;
-    z-index: 2;
-}
-
-.projects {
-    height: 100%;
-    transition: 1.5s cubic-bezier(0,0,0.25,1);
-}
-
-.project-item {
-    display: flex;
-    height: 100%;
-    justify-content: center;
-    align-items: center;
-    background-position: center !important;
-    background-size: cover !important;
-    background-repeat: no-repeat !important;
-    font-size: 100px;
-    color: white;
-    transform: skew(0);
-}
-
-.project-item-logo {
-    display: flex;
-    justify-content: center;
-    background-position: center center;
-    transition: 1.5s ease-out;
-}
-
-.project-item-logo img {
-    width: 60%;
-}
-
-.slider-navigation-container {
-    position: absolute;
-    left: 15px;
-    top: 50%;
-    transform: translateY(-50%);
-}
-
-.slider-navigation-item-wrapper {
-    padding: 13px 0;
-    cursor: pointer;
-}
-
-.slider-navigation-item {
-    width: 17px;
-    height: 3px;
-    background: #F7F7F8;
-    border-radius: 10px;
-    transition: .3s;
-}
-
-.slider-navigation-item-wrapper:hover .slider-navigation-item {
-    width: 25px;
-    background: #FB4747;
-}
-
-.slider-navigation-item-active {
-    width: 25px;
-    background: #FB4747;
-}
-
-.contact-button {
-    margin-top: 10px;
-    color: #F7F7F8;
-    transform: rotate(-180deg);
-    transform-origin: center center;
-    writing-mode: vertical-rl;
-    text-orientation: mixed;
-    cursor: pointer;
-    transition: .3s;
-}
-
-.contact-button:hover {
-    color: #FB4747;
-}
-
-.more-projects {
-    display: flex;
-    align-items: center;
-    position: absolute;
-    bottom: 15px;
-    left: 15px;
-    color: white;
-    font-size: 13px;
-    cursor: pointer;
-}
-
-.more-projects img {
-    margin-right: 15px;
-}
-
-.project-item-preview {
-    position: relative;
-    background: #340f0f;
-}
-
-.project-item-preview-input {
-    display: none;
-}
-
-.project-item-preview-input-ui {
-    position: absolute;
-    left: 50%;
-    top: 30px;
-    font-size: 12px;
-    padding: 5px 10px;
-    background: #023cff;
-    border-radius: 2px;
-    box-shadow: 0 5px 10px 0 rgba(0,0,0,0.4);
-    transform: translateX(-50%);
-    transition: .5s;
-    cursor: pointer;
-}
-
-.project-item-preview-input-ui:hover {
-    box-shadow: 0 10px 10px 0 rgba(0,0,0,0.8);
-}
