--- src/components/ProjectPresentation/ProjectPresentation.css
@@ -1,77 +1,0 @@
-.samples-container {
-    position: relative;
-    width: 100%;
-    height: 100%;
-}
-
-.samples {
-    position: relative;
-    width: 100%;
-    height: 100%;
-    transition: 1s ease-out;
-}
-
-.samples-wrapper {
-    position: relative;
-    flex-wrap: nowrap;
-    width: 100%;
-    min-height: 100%;
-    transition: 1s;
-}
-
-.samples-item {
-    position: absolute;
-    min-width: 100%;
-    height: 100%;
-    background-position: center !important;
-    background-size: cover !important;
-    background-repeat: no-repeat !important;
-    transition: 1s ease-out;
-}
-
-.navigation-arrow {
-    position: absolute;
-    top: 50%;
-    transform: translateY(-50%);
-    z-index: 5;
-    cursor: pointer;
-}
-
-.navigation-arrow img {
-    width: 30px;
-}
-
-.left-navigation-arrow {
-    left: 20px;
-}
-
-.right-navigation-arrow {
-    right: 20px;
-}
-
-
-.my-node {
-    position: absolute;
-    height: 100%;
-    width: 100%;
-    overflow: hidden;
-}
-
-
-/*TODO: rename 'my-node' class*/
-.my-node.enter {
-    opacity: 0;
-}
-.my-node.enter-done {
-    opacity: 1;
-    transition: 1000ms;
-}
-.my-node.exit {
-    opacity: 1;
-    transition: 1000ms;
-}
-.my-node.exit-done {
-    opacity: 0;
-    transition: 1000ms;
-}
-
