--- src/components/Projects/Projects.css
@@ -1,14 +1,0 @@
-.container {
-    display: flex;
-    flex-wrap: wrap;
-    padding: 10px 0;
-}
-
-.projects-item {
-    flex-basis: 25%;
-    overflow: hidden;
-}
-
-.projects-item img {
-    width: 100%;
-}
