--- src/enums/ProjectMode.js
@@ -1,6 +1,0 @@
-const ProjectMode = {
-    ProjectToAdd: 'ProjectToAdd',
-    ProjectToEdit: 'ProjectToEdit',
-};
-
-export default ProjectMode;
