--- src/helpers/defaultData.js
@@ -1,15 +1,0 @@
-class LogoPreview {
-    constructor(mode = null) {
-        this.name = '';
-        this.logoSrc = null;
-        this.description = '';
-        this.logoCreatedAt = '';
-        this.webSite = '';
-        this.rightTopSectionBg = 'black';
-        this.rightTopSectionFontColor = 'white';
-        this.samplesProject = [];
-        this.mode = mode;
-    }
-}
-
-export { LogoPreview };
