--- src/helpers/helpers.js
@@ -1,11 +1,0 @@
-const fileToBase64 = file => new Promise((resolve, reject) => {
-    if (!file) {
-        reject('');
-    }
-    const reader = new FileReader();
-    reader.readAsDataURL(file);
-    reader.onload = () => resolve(reader.result);
-    reader.onerror = error => reject(error);
-});
-
-export { fileToBase64 };
