+++ src/images/works/index.js
@@ -1,0 +1,39 @@
+import beard from './01_beardstyle_800x900.jpg';
+import prospervideo from './02_prospervideo_800x900.jpg';
+import ditchingdisponsable from './03_ditchingdisponsable_800x900.jpg';
+import emilyjrealty from './04_emilyjrealty_800x900.jpg';
+import igp from './05_igp_800x900.jpg';
+import lemontnoir from './06_lemontnoir_800x900.jpg';
+import sakseide from './07_sakseide_800x900.jpg';
+import mainevent from './08_mainevent_800x900.jpg';
+import aaronanthony from './09_aaronanthony_800x900.jpg';
+import britishloghomes from './10_britishloghomes_800x900.jpg';
+import starsandbones from './11_starsandbones_800x900.jpg';
+import archx from './12_archx_800x900.jpg';
+
+export const works = [
+    [
+        { avatarUrl: beard, name: 'BEARD STYLE', description: 'Helping men grow and groom awesome beards' },
+        { avatarUrl: prospervideo, name: 'PROSPER.VIDEO', description: 'Mobile-first video production studio' },
+    ],
+    [
+        { avatarUrl: ditchingdisponsable, name: 'DITCHING DISPONSABLES', description: 'You make a difference' },
+        { avatarUrl: emilyjrealty, name: 'EMILY J', description: 'Real estate company' },
+    ],
+    [
+        { avatarUrl: igp, name: 'INSTANT GLOBAL PAYMENTS', description: 'Send money to any country with world-class exchange rates' },
+        { avatarUrl: lemontnoir, name: 'LE MONTNOIR', description: 'Fine furniture' },
+    ],
+    [
+        { avatarUrl: sakseide, name: 'SAKSEIDE CONSULTING', description: 'Change management, organizational psychology' },
+        { avatarUrl: mainevent, name: 'MAIN EVENT CATERING', description: 'Voted Anchorage, “Alaska’s Best Caterer”' },
+    ],
+    [
+        { avatarUrl: aaronanthony, name: 'AARON ANTHONY', description: '' },
+        { avatarUrl: britishloghomes, name: 'BRITISH LOG HOMES', description: 'The home of handcrafted log cabins in the UK' },
+    ],
+    [
+        { avatarUrl: starsandbones, name: 'STARS&BONES', description: 'Audiovisual production company' },
+        { avatarUrl: archx, name: 'ARCH X', description: 'Residential architectural design' },
+    ],
+];
