+++ src/index.js
@@ -1,16 +1,11 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { Provider } from 'react-redux';
 import './index.css';
 import App from './App';
 import * as serviceWorker from './serviceWorker';
 
-import store from "./store";
-
 ReactDOM.render(
-    <Provider store={store}>
-        <App />
-    </Provider>
+    <App />
     ,
     document.getElementById('root')
 );
