--- src/pages/Login/Login.css
@@ -1,56 +1,0 @@
-.login-container {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    width: 100vw;
-    height: 100vh;
-    background: #FB4747;
-}
-
-.login-form {
-    width: 30%;
-    padding: 20px 35px;
-    background: white;
-    box-shadow: 0px 5px 10px 0px rgba(0,0,0,0.4);
-}
-
-.control-field {
-    display: block;
-    width: 100%;
-    margin: 20px 0;
-    padding: 15px;
-    background: #F2F2F2;
-    font-size: 16px;
-    box-sizing: border-box;
-    border: none;
-}
-
-.control-field:focus {
-    outline: none;
-}
-
-.form-submit-button {
-    display: block;
-    width: 100%;
-    margin: 0;
-    padding: 15px 0;
-    background: #FB4747;
-    color: white;
-    text-align: center;
-    border: none;
-    cursor: pointer;
-}
-
-.form-submit-button:focus {
-    outline: none;
-}
-
-.form-submit-button:disabled {
-    opacity: .5;
-    cursor: default;
-}
-
-.error-container {
-    margin-top: 10px;
-    color: red;
-}
